[accel-record] export deepMerge() function
