+++ .changeset/swift-experts-taste.md
@@ -1,0 +1,6 @@
+---
+"accel-record-core": minor
+"accel-record": minor
+---
+
+export deepMerge() function
