+++ packages/accel-record/src/merge.ts
@@ -1,0 +1,1 @@
+export * from "accel-record-core/dist/merge.js";
